fix(playground): fix arrow icon does not work correctly on webkit
